submit

+++ libft/ft_itoa.c
@@ -6,7 +6,7 @@
 /*   By: mosada <mosada@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/29 16:18:54 by mosada            #+#    #+#             */
-/*   Updated: 2023/10/07 17:41:26 by mosada           ###   ########.fr       */
+/*   Updated: 2023/10/12 21:21:25 by mosada           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ static int	intcount(long long n)
 char	*ft_itoa(int n)
 {
 	char		*result;
-	int			count;
+	size_t		count;
 	bool		is_negative;
 	long long	num;
