+++ libft/ft_strtrim.c
@@ -6,7 +6,7 @@
 /*   By: mosada <mosada@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/29 11:19:05 by mosada            #+#    #+#             */
-/*   Updated: 2023/10/07 15:18:04 by mosada           ###   ########.fr       */
+/*   Updated: 2023/10/12 21:19:19 by mosada           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ char	*ft_strtrim(char const *s1, char const *set)
 {
 	char	*result;
 	int		i;
-	int		count;
+	size_t	count;
 	int		size;
 
 	if (!s1 || !set)
